requirements.txt
# CuratAI - Dependencies

# Core Python (3.9+)

# AI & LLM
openai>=0.27.0
anthropic>=0.7.0

# Opik - Orchestration & Monitoring
opik>=0.1.0

# Web3 & Blockchain
web3>=6.0.0
py-ipfs-http-client>=0.8.0
didkit>=0.3.0

# Data Processing
pandas>=1.3.0
numpy>=1.21.0

# NLP & Text Processing
nltk>=3.6.0

# Configuration
pyyaml>=6.0
python-dotenv>=0.19.0

# Database (optional, for persistence)
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0  # PostgreSQL

# Utilities
requests>=2.28.0
click>=8.0.0  # CLI
tqdm>=4.62.0  # Progress bars

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0
unittest-mock>=1.4.0

# Development
black>=22.0.0  # Code formatting
flake8>=4.0.0  # Linting
mypy>=0.950  # Type checking
isort>=5.10.0  # Import sorting

# Documentation
sphinx>=4.5.0
sphinx-rtd-theme>=1.0.0

# Logging & Monitoring
python-json-logger>=2.0.0

# Optional: Jupyter for interactive demos
jupyter>=1.0.0
jupyterlab>=3.0.0
ipython>=7.0.0
